Asynchronous initialization of shader cache.
Put the shader initialization logic into a separate thread and lock it
to reduce the load on RenderThread. The interval between initialization
and use of ShaderCache is a long time (~50ms). This period can be enough
for ShaderCache to execute the initialization logic.

Due to the protection of mMutex, there will be no thread safety issues.

Change-Id: I7fb3ab06d0003eeba1a542d872177ab80e7d4d73
Signed-off-by: Liangliang Sui <coolsui.coding@gmail.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>